+++ crates/cubecl-llvm/src/cpu/entrypoint.rs
@@ -13,7 +13,7 @@ use pliron::identifier::Identifier;
 use pliron::linked_list::ContainsLinkedList;
 use pliron_llvm::types::PointerType as LlvmPointerType;
 
-use crate::shared::polyfill::synchronization::ATTR_SYNC_CUBE_STATE;
+use crate::cpu::synchronization::ATTR_SYNC_CUBE_STATE;
 use crate::shared::shared_memory::declares_shared_memory;
 
 pub const CPU_RUNTIME_BUILTINS: [Builtin; 6] = [
