+++ crates/cubecl-llvm/src/cpu/jit/engine.rs
@@ -11,7 +11,7 @@ use pliron_llvm::llvm_sys::target::initialize_native;
 use pliron_llvm::to_llvm_ir;
 
 use super::data::PlironData;
-use crate::shared::shared_memory::SharedMemories;
+use crate::cpu::shared_memory::SharedMemories;
 
 /// Host ABI of a JIT'd kernel: `(buffer_ptrs, cube_count_x/y/z, unit_pos_x/y/z, sync_cube_state,
 /// metadata)`. The variable-count pointers — the buffers and then the shared memories — are
