+++ crates/cubecl-llvm/src/lib.rs
@@ -4,8 +4,12 @@ extern crate derive_new;
 extern crate alloc;
 
 pub mod amdgpu;
+pub mod cpu;
 pub mod shared;
+pub mod target;
 
-pub use shared::jit::data::{PlironData, SharedData};
-pub use shared::jit::engine::{KernelRequirements, PlironEngine};
-pub use shared::{AmdGpuModule, LlvmTarget, PlironArtifact, PlironCompiler, PlironOptions};
+pub use cpu::jit::data::{PlironData, SharedData};
+pub use cpu::jit::engine::{KernelRequirements, PlironEngine};
+pub use cpu::shared_memory::SharedMemories;
+pub use shared::{AmdGpuModule, PlironArtifact, PlironCompiler, PlironOptions};
+pub use target::LlvmTarget;
