+++ crates/cubecl-llvm/src/shared/metadata.rs
@@ -1,6 +1,5 @@
-//! Rewrites the kernel entry ABI to what the JIT host calls.
+//! Rewrites the kernel entry ABI to what the host calls.
 
-use core::cell::RefCell;
 use cubecl_core::ir::attributes::{
     ATTR_BUFFER_BINDING, BufferBindingAttr, FuncInterface, IndexAttr,
 };
@@ -18,9 +17,8 @@ use pliron::basic_block::BasicBlock;
 use pliron::builtin::attributes::TypeAttr;
 use pliron::builtin::ops::{ConstantOp, FuncOp};
 use pliron::builtin::types::FunctionType;
-use std::rc::Rc;
 
-use crate::shared::shared_memory::{SharedDeclarations, SharedMemories};
+use crate::shared::shared_memory::SharedDeclarations;
 use crate::shared::to_llvm::ty::cube_type_to_llvm;
 
 /// `(op, buffer_idx, result)` for each `cube.buffer_len`, gathered during the walk so the ops
@@ -62,19 +60,6 @@ pub struct LowerEntryAbiPass {
 impl LowerEntryAbiPass {
     pub fn new(info: Info, layout: Box<dyn EntryArgLayout>) -> Self {
         Self { info, layout }
-    }
-}
-
-/// The CPU entry layout: buffers and shared memories are collapsed behind one `%buffer_ptrs`
-/// indirection table, which is how the JIT host hands the kernel its resources.
-pub struct TableArgs {
-    /// Filled in with the shared memory the host must reserve, see [`SharedMemories`].
-    shared_memories: Rc<RefCell<SharedMemories>>,
-}
-
-impl TableArgs {
-    pub fn new(shared_memories: Rc<RefCell<SharedMemories>>) -> Self {
-        Self { shared_memories }
     }
 }
 
@@ -218,54 +203,6 @@ impl Pass for LowerEntryAbiPass {
     }
 }
 
-impl EntryArgLayout for TableArgs {
-    fn present_args(
-        &self,
-        ctx: &mut Context,
-        func: FuncOp,
-        buffers: &[(usize, usize, Value)],
-        shared: SharedDeclarations,
-    ) {
-        let entry = func.get_entry_block(ctx);
-
-        let shared_base = (buffers.iter())
-            .map(|(_, buffer_pos, _)| buffer_pos + 1)
-            .max()
-            .unwrap_or(0);
-        if !buffers.is_empty() || !shared.is_empty() {
-            let table_ty = table_ty(ctx);
-            BasicBlock::insert_argument(entry, ctx, 0, table_ty);
-            let buffer_ptrs = entry.deref(ctx).get_argument(0);
-            let terminator = entry
-                .deref(ctx)
-                .get_terminator(ctx)
-                .expect("entry block must be terminated");
-
-            for (_idx, buffer_pos, old_val) in buffers.iter() {
-                let buffer_ty = old_val.get_type(ctx);
-                let buffer = load_table(ctx, buffer_ptrs, *buffer_pos, buffer_ty, terminator);
-                old_val.replace_all_uses_with(ctx, &buffer);
-            }
-
-            let blocks = shared.lower(ctx, buffer_ptrs, shared_base, terminator);
-            if !blocks.is_empty() {
-                *self.shared_memories.borrow_mut() = SharedMemories {
-                    base: shared_base,
-                    blocks,
-                };
-            }
-
-            let mut removed: Vec<usize> = buffers.iter().map(|(i, _, _)| i + 1).collect();
-            removed.sort_unstable();
-            for idx in removed.into_iter().rev() {
-                BasicBlock::remove_argument(entry, ctx, idx);
-            }
-        }
-
-        rebuild_func_type(ctx, func);
-    }
-}
-
 /// Rebuilds `func`'s `FunctionType` from its entry block's current argument types. Called once
 /// [`EntryArgLayout::present_args`] has finished changing the arguments, since that is target
 /// specific but this reconciliation of the declared signature is not.
@@ -298,7 +235,7 @@ fn array_ptr_to(ctx: &Context, elem: TypeHandle) -> TypeHandle {
 
 /// Type of `%buffer_ptrs`, the table holding one pointer per buffer and then per shared memory.
 /// What each of them points at is up to the slot's owner, so the table holds them opaquely.
-fn table_ty(ctx: &Context) -> TypeHandle {
+pub(crate) fn table_ty(ctx: &Context) -> TypeHandle {
     array_ptr_to(ctx, ptr_to(ctx, BytesType::get(ctx).into()))
 }
