+++ crates/cubecl-llvm/src/shared/polyfill/mod.rs
@@ -1,7 +1,5 @@
 pub mod math;
-pub mod ordered_atomic;
-pub mod synchronization;
-mod transcendental;
+pub mod transcendental;
 
 use cubecl_core::ir::{NamedRewrite, prelude::*};
