+++ crates/cubecl-llvm/src/shared/polyfill/transcendental/mod.rs
@@ -2,13 +2,13 @@
 //! vector intrinsics: `llvm.exp` on a vector reaches codegen with nothing but a declare
 //! behind it, and every lane becomes a libm call that a dozen fused multiply-adds beat.
 
-mod base;
+pub(crate) mod base;
 mod exponential;
 mod hyperbolic;
 mod logarithm;
 mod trigonometry;
 
-pub(super) use exponential::exp;
-pub(super) use hyperbolic::tanh;
-pub(super) use logarithm::ln;
-pub(super) use trigonometry::{cos, sin};
+pub use exponential::exp;
+pub use hyperbolic::tanh;
+pub use logarithm::ln;
+pub use trigonometry::{cos, sin};
