+++ crates/cubecl-llvm/src/shared/shared_memory.rs
@@ -1,35 +1,20 @@
-//! Lowering of the shared `memory.declare_variable` to a pointer the host provides.
+//! What a kernel declares as shared memory, as both targets have to see it.
 //!
-//! Every unit of a cube runs the kernel on a thread of its own, so shared memory cannot be a
-//! stack allocation: all the units have to see the same bytes. The host reserves a block per
-//! shared memory instead (out of the stream's shared memory pool) and passes it the same way as a
-//! buffer, in the pointer table the entry ABI hands to the kernel. Declaring shared memory then
-//! costs no more than reading a kernel argument.
-//!
-//! Like the units, the cubes of a launch share those blocks, which is what a cube barrier is for:
-//! without one, a unit racing ahead to the next cube may overwrite what the units still finishing
-//! the current one are reading.
+//! Collecting the declarations and measuring the block each one needs is target independent.
+//! What replaces a declaration is not, and belongs to the target's
+//! [`EntryArgLayout`](super::metadata::EntryArgLayout): see
+//! [`cpu::shared_memory`](crate::cpu::shared_memory) for how the CPU host provides its blocks.
 
 use cubecl_core::ir::AddressSpace;
 use cubecl_core::ir::dialect::memory::DeclareVariableOp;
 use cubecl_core::ir::interfaces::SizedType;
 use cubecl_core::ir::prelude::*;
 
-use crate::shared::metadata::load_table;
-
 /// A block of shared memory the host must reserve to launch the kernel.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub struct SharedMemoryBlock {
     pub size: usize,
     pub align: usize,
-}
-
-/// The shared memory of a kernel: the blocks to reserve, and the slot of the pointer table their
-/// pointers go to. They sit right after the buffers, which own the slots before `base`.
-#[derive(Clone, Debug, Default)]
-pub struct SharedMemories {
-    pub base: usize,
-    pub blocks: Vec<SharedMemoryBlock>,
 }
 
 /// Whether `op` declares any shared memory, i.e. whether its cube iterations share state that
@@ -45,7 +30,7 @@ pub fn declares_shared_memory(ctx: &Context, op: Ptr<Operation>) -> bool {
 /// `(op, result, block)` for each shared declaration, gathered during the walk so the ops can be
 /// rewritten once the walker no longer holds them borrowed.
 #[derive(Default)]
-pub struct SharedDeclarations(Vec<(Ptr<Operation>, Value, SharedMemoryBlock)>);
+pub struct SharedDeclarations(pub(crate) Vec<(Ptr<Operation>, Value, SharedMemoryBlock)>);
 
 impl SharedDeclarations {
     /// Collects every shared memory declared under `root`.
@@ -80,28 +65,5 @@ impl SharedDeclarations {
 
     pub fn is_empty(&self) -> bool {
         self.0.is_empty()
-    }
-
-    /// Replaces every declaration by the pointer `table` holds for it, starting at slot `base`.
-    /// The loads go before `before`, which must be in a block dominating every use.
-    pub fn lower(
-        self,
-        ctx: &mut Context,
-        table: Value,
-        base: usize,
-        before: Ptr<Operation>,
-    ) -> Vec<SharedMemoryBlock> {
-        self.0
-            .into_iter()
-            .enumerate()
-            .map(|(offset, (decl, result, block))| {
-                let ptr_ty = result.get_type(ctx);
-                let ptr = load_table(ctx, table, base + offset, ptr_ty, before);
-
-                result.replace_all_uses_with(ctx, &ptr);
-                Operation::erase(decl, ctx);
-                block
-            })
-            .collect()
     }
 }
